Refactored database create client code into it's own file
